# [no squash] Fix network window size logic