this is supposed to fix #14765

## How to test

1. join a busy server with 5.9.0 or older.
2. update a busy server to this PR.
3. try to reproduce the issue by joining from a poor connection.