@sfan5 2024-09-02
> it might need more adjustments to allow the window to grow faster to not regress media transfer speed too much.
> (testing needed to determine if this is the case)

I tested a bit with a latency of 126ms over the internet.
the results:
|           | 6MB  | 16MB |
|-----------|------|------|
| before PR | 10s  | 13s  |
| with PR   | 11s  | 15s  |
| wget      | 1,5s | 2,2s |

Obviously a smaller window means more transfer time, but the difference is less than I expected. So that doesn't need adjustment.
wget is shown for comparison (our protocol implementation is really terrible).